requirements.dev.txt
# dev dependencies
ipython>=8.0.0,<=8.25.0
ipdb
pudb
pytest-watch

# Install project as editable
-e .
